Infobeansblogs/requirements.txt
alembic==1.7.7
alipay==0.7.4
aniso8601==9.0.1
arrow==1.2.3
async-timeout==4.0.2
Babel==2.11.0
bcrypt==4.0.1
blinker==1.5
certifi==2021.5.30
charset-normalizer==2.0.12
click==8.0.4
dataclasses==0.8
dnspython==2.2.1
dominate==2.8.0
email-validator==1.3.1
Flask==2.0.3
Flask-Babel==2.0.0
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-Dance==7.0.0
Flask-DebugToolbar==0.13.1
Flask-JWT==0.3.2
Flask-Login==0.5.0
Flask-Migrate==4.0.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==2.5.1
Flask-TinyMCE==1.0.0
Flask-WTF==1.0.1
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
oauthlib==3.2.2
packaging==21.3
Pillow==8.4.0
pluggy==1.0.0
PyJWT==1.4.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
redis==4.3.6
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==1.4.49
typing_extensions==4.1.1
urllib3==1.26.16
URLObject==2.4.3
visitor==0.1.3
Werkzeug==2.0.3
WTForms==3.0.0
zipp==3.6.0
